requirements.txt
Package                  Version
------------------------ ------------
accelerate               0.32.1
aiohttp                  3.9.5
aiosignal                1.3.1
asttokens                2.4.1
async-timeout            4.0.3
attrs                    23.2.0
audioread                3.0.1
backcall                 0.2.0
certifi                  2024.6.2
cffi                     1.16.0
charset-normalizer       3.3.2
click                    8.1.7
comm                     0.2.2
contourpy                1.1.1
cycler                   0.12.1
datasets                 2.20.0
debugpy                  1.8.2
decorator                5.1.1
diffusers                0.30.0.dev0
dill                     0.3.8
docker-pycreds           0.4.0
executing                2.0.1
filelock                 3.13.1
fonttools                4.53.0
frozenlist               1.4.1
fsspec                   2024.2.0
gitdb                    4.0.11
GitPython                3.1.43
huggingface-hub          0.23.4
idna                     3.7
importlib_metadata       8.0.0
importlib_resources      6.4.0
ipykernel                6.29.5
ipython                  8.12.0
jedi                     0.19.1
Jinja2                   3.1.3
joblib                   1.4.2
jupyter_client           8.6.2
jupyter_core             5.7.2
kiwisolver               1.4.5
lazy_loader              0.4
librosa                  0.10.2.post1
llvmlite                 0.41.1
MarkupSafe               2.1.5
matplotlib               3.7.5
matplotlib-inline        0.1.7
mpmath                   1.3.0
msgpack                  1.0.8
multidict                6.0.5
multiprocess             0.70.16
nest_asyncio             1.6.0
networkx                 3.0
numba                    0.58.1
numpy                    1.24.1
nvidia-cublas-cu11       11.11.3.6
nvidia-cuda-cupti-cu11   11.8.87
nvidia-cuda-nvrtc-cu11   11.8.89
nvidia-cuda-runtime-cu11 11.8.89
nvidia-cudnn-cu11        8.7.0.84
nvidia-cufft-cu11        10.9.0.58
nvidia-curand-cu11       10.3.0.86
nvidia-cusolver-cu11     11.4.1.48
nvidia-cusparse-cu11     11.7.5.86
nvidia-nccl-cu11         2.20.5
nvidia-nvtx-cu11         11.8.86
opencv-python            4.10.0.84
packaging                24.1
pandas                   2.0.3
parso                    0.8.4
pexpect                  4.9.0
pickleshare              0.7.5
pillow                   10.2.0
pip                      24.0
platformdirs             4.2.2
pooch                    1.8.2
prompt_toolkit           3.0.47
protobuf                 5.27.2
psutil                   6.0.0
ptyprocess               0.7.0
pure-eval                0.2.2
pyarrow                  16.1.0
pyarrow-hotfix           0.6
pycparser                2.22
Pygments                 2.18.0
pyparsing                3.1.2
python-dateutil          2.9.0
pytz                     2024.1
PyYAML                   6.0.1
pyzmq                    26.0.3
regex                    2024.5.15
requests                 2.32.3
safetensors              0.4.3
scikit-learn             1.3.2
scipy                    1.10.1
seaborn                  0.13.2
sentry-sdk               2.10.0
setproctitle             1.3.3
setuptools               70.1.1
six                      1.16.0
smmap                    5.0.1
soundfile                0.12.1
soxr                     0.3.7
stack-data               0.6.2
sympy                    1.12
threadpoolctl            3.5.0
tokenizers               0.19.1
torch                    2.3.0+cu118
torchaudio               2.3.0+cu118
torchvision              0.18.0+cu118
tornado                  6.4.1
tqdm                     4.66.4
traitlets                5.14.3
transformers             4.42.3
triton                   2.3.0
typing_extensions        4.12.2
tzdata                   2024.1
urllib3                  1.26.19
wandb                    0.17.4
wcwidth                  0.2.13
wheel                    0.43.0
xxhash                   3.4.1
yarl                     1.9.4
zipp                     3.19.2